Sprinkle (long long) casts where %lld is being used to print daddr_t variables. Some random whitespace/knf repairs encountered on the way.
ok miod@ on inspection, feedback & more suggestions from millert@

+++ sbin/fsck_ffs/inode.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: inode.c,v 1.38 2013/06/11 16:42:04 deraadt Exp $	*/
+/*	$OpenBSD: inode.c,v 1.39 2013/11/01 17:36:18 krw Exp $	*/
 /*	$NetBSD: inode.c,v 1.23 1996/10/11 20:15:47 thorpej Exp $	*/
 
 /*
@@ -240,7 +240,8 @@ chkrange(daddr_t blk, int cnt)
 	    fragnum(&sblock, blk) + cnt > sblock.fs_frag) {
 		if (debug)
 			printf("bad size: blk %lld, offset %lld, size %d\n",
-			    blk, fragnum(&sblock, blk), cnt);
+			    (long long)blk, fragnum(&sblock, blk),
+			    cnt);
 		return (1);
 	}
 	c = dtog(&sblock, blk);
@@ -248,9 +249,10 @@ chkrange(daddr_t blk, int cnt)
 		if ((blk + cnt) > cgsblock(&sblock, c)) {
 			if (debug) {
 				printf("blk %lld < cgdmin %lld;",
-				    blk, cgdmin(&sblock, c));
+				    (long long)blk, cgdmin(&sblock, c));
 				printf(" blk + cnt %lld > cgsbase %lld\n",
-				    blk + cnt, cgsblock(&sblock, c));
+				    (long long)(blk + cnt),
+				    cgsblock(&sblock, c));
 			}
 			return (1);
 		}
@@ -258,9 +260,9 @@ chkrange(daddr_t blk, int cnt)
 		if ((blk + cnt) > cgbase(&sblock, c+1)) {
 			if (debug)  {
 				printf("blk %lld >= cgdmin %lld;",
-				    blk, cgdmin(&sblock, c));
+				    (long long)blk, cgdmin(&sblock, c));
 				printf(" blk + cnt %lld > sblock.fs_fpg %d\n",
-				    blk+cnt, sblock.fs_fpg);
+				    (long long)(blk+cnt), sblock.fs_fpg);
 			}
 			return (1);
 		}
