+++ sys/dev/softraid_raid1.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: softraid_raid1.c,v 1.53 2013/06/11 16:42:13 deraadt Exp $ */
+/* $OpenBSD: softraid_raid1.c,v 1.54 2013/11/01 17:36:19 krw Exp $ */
 /*
  * Copyright (c) 2007 Marco Peereboom <marco@peereboom.us>
  *
@@ -426,7 +426,7 @@ sr_raid1_wu_done(struct sr_workunit *wu)
 	/* If all I/O failed, retry reads and give up on writes. */
 	if (xs->flags & SCSI_DATA_IN) {
 		printf("%s: retrying read on block %lld\n",
-		    sd->sd_meta->ssd_devname, wu->swu_blk_start);
+		    sd->sd_meta->ssd_devname, (long long)wu->swu_blk_start);
 		if (wu->swu_cb_active == 1)
 			panic("%s: sr_raid1_intr_cb",
 			    DEVNAME(sd->sd_sc));
@@ -436,7 +436,7 @@ sr_raid1_wu_done(struct sr_workunit *wu)
 			return SR_WU_RESTART;
 	} else {
 		printf("%s: permanently failing write on block %lld\n",
-		    sd->sd_meta->ssd_devname, wu->swu_blk_start);
+		    sd->sd_meta->ssd_devname, (long long)wu->swu_blk_start);
 	}
 
 	wu->swu_state = SR_WU_FAILED;
